Refactoring the basic framework
