+++ rllm/preprocessing/__init__.py
@@ -18,12 +18,12 @@ from rllm.preprocessing._text_tokenize import (
     standardize_tokenizer_output,
     tokenize_merged_cols,
     save_column_name_tokens,
-    TransTabDataExtractor,
 )
 from rllm.preprocessing._word_embedding import (
     TextEmbedderConfig,
     embed_text_column,
 )
+from rllm.preprocessing.transtab_preprocessing import TransTabDataExtractor
 
 __all__ = [
     # fillna
